Added rest toolbar.  Not sure about this; I originally tried it in the same toolbar as the notes (which really makes more sense) but that looked pretty bizarre too.  Maybe we need special pixmaps for the toolbars, with more regular dimensions than the ones we use on the staff.  Maybe we need a RG2.1-style palette, although RG2.1's was pretty unwieldy.
Rest insertion is currently incorrect; the code for adding notes
to chords is being followed, but of course a rest with the same
absolute time as a note is not considered part of a chord so the
display is as if an actual insertion was taking place.

Also added breves to the toolbars, and breve support to notepixmapfactory.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@486 cf6a83e7-f017-4305-b0b3-678bde24f230