Considerable modifications to Composition, NotationView, Notation[VH]Layout, Staff etc., in order to make the NotationView capable of representing more than one Track at once.  (It can't draw them yet, but it can hold them.)
Composition now has a time reference track in addition to the other tracks;
this is used to keep time sig changes in sync between other tracks.

NotationView now holds a set of Staffs, and the Staff now holds the
ViewElementsManager, NotationElementList and so on.  This seemed like a
reasonably logical breakdown.  Notation[VH]Layout mildly refactored to
take Staff args, but HLayout needs quite a bit more work yet.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@903 cf6a83e7-f017-4305-b0b3-678bde24f230