Moved BasicSelectionCommand into basiccommand.h (just as EventSelections are intended to be generic and available in any view, so should this be)
Made MultiViewCommandHistory's commandExecuted signal take a KCommand*
argument, so now each EditView can attach to it, receive the signal,
get the Segment and extents from the given command (if it's a BasicCommand),
and invoke a refresh.

Side-effect: added undo/redo and automatic refresh to Matrix view, but it
won't work with the built-in Matrix editing operations until they're made
into commands.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@1435 cf6a83e7-f017-4305-b0b3-678bde24f230