Farily large refactoring to move segment tools in a toolbox like in Edit Views. - EditTool and SegmentTool now derive from BaseTool which handles the menu creation and "toolbox compliance"  (ready()/stow() methods) - Creation of SegmentToolBox which holds all the tools - EditToolBox and SegmentToolBox derive from BaseToolBox
The menu creation process is different between EditTools and SegmentTools, because
EditTools are creating their own KActions, and therefore need to specify another xml
file and create the menu from it, while SegmentTools create a menu made of already
existing actions.

Handling of selection when right-clicking, which is the whole purpose of all this, is
ironically not there yet.

the


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@3916 cf6a83e7-f017-4305-b0b3-678bde24f230