* some more menu shuffling * makes the panic button stop playback, and provides a hopefully more informative   status message to make it more obvious that the progess bar doesn't mean the   events have all *fnished* being transmitted
NOTE:  I never could figure out how to get the default Toolbar and Statusbar entries off the Settings menu and onto the submenus for notation/matrix views.  Some piece to the puzzle I'm just not getting.  The current effect is pretty stupid looking, and if this can't be fixed it would probably be better to revert entirely.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@5646 cf6a83e7-f017-4305-b0b3-678bde24f230