Fix #1538844 by using a better method from Pitch
Even better verbose error messages.

Back out broken hack for #1538848, and comment out the errant entries in the
data instead.  (New version of data not committed yet, because I'm still
sorting another bug.)

A broken fix in notationcommands.cpp made it in here too.  I'm still sorting
that out also, and this code needs gutting again, but this version is only
broken in a different way from what was already completely broken, so I'm not
going to bother backing it out of this commit.



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@7438 cf6a83e7-f017-4305-b0b3-678bde24f230