Merge all the way back to rev 8390, which backs out all of Heikki's refactoring, and Arnout's change to use a different version of some other class, but yields a working pitch chooser dialog with no apparent ill side effects, in spite of all the ways one might expect this to fail.
The signals and slots seem to be frigged up somehow in the 8391 version.  Maybe
it's obvious to someone else, and easy to fix that version, and put Heikki's
refactoring back into action, but I spent hours on it without puzzling it out,
and once I saw that it was possible to simply go back and grab the 8390
revision and merge it straight into 8586 verbatim, I opted for the quick fix.
I'm sure whatever Heikki was doing had merit, but it wasn't worth the price of
breaking the drag and mouse wheel functions of this dialog, which then made it
into two releases to our continuing embarrassment.

He should feel free to start with the 8586 revision and fix the problem if he so
chooses.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@8587 cf6a83e7-f017-4305-b0b3-678bde24f230