* Move MultiViewCommandHistory to plain CommandHistory, and properly   handle its singleton-nature * Finish up AudioMixerWindow actions
I've added some logic to ActionFileParser to make it add the undo/redo
actions found in the CommandHistory to the current window if they are
declared in the rc file, even if they have not been defined in the
subclass itself.  This may mean that subclasses of ActionFileParser
that have proper rc files with undo/redo declared in them will no
longer need any equivalent of the (old-style) MultiViewCommandHistory::
attachView() or the (new-style) CommandHistory::registerMenu() or
registerToolbar().



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/qt4@9626 cf6a83e7-f017-4305-b0b3-678bde24f230