* fix Transport blinkiness
* fix selection background color, with caveats

* update stylesheet

NOTES:

The selected track background color works now when no stylesheet is loaded, but
loading the stylesheet causes problems in this area.

I'm not sure how to resolve all of this, really.  I converted the Qt3 code to
Qt4, but there are caveats everywhere about using this kind of code with
stylesheets.  The problem is the colors won't pull from the stylesheet, because
these labels for the selecte track aren't really in a "selected" state.  We were
creating the appearance of being "selected" artificially, as far as I can tell.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/qt4@9832 cf6a83e7-f017-4305-b0b3-678bde24f230