* fix the file dialogs (that was refreshingly easy!!)
* comment out the color configuration page so it is never seen

I've been looking into the color configuration bits, and beyond mere style
issues, these are quite thoroughly busted.  I don't want to take the approach of
just tossing something aside merely because it is challenging to fix, but in
this case I think we may well be justified in removing the color table editor.
It was rarely, if ever used, and it is very likely that anybody who actually
cares about segment colors already has all the colors they could ever want
pre-defined since I added the gigantic (400+) list of named colors to the
default studio some years ago.  This isn't a drawing application, and people
won't likely care if they can't have the exact shade of red they want.

In fact, I added the gigantic list of pre-created, pre-named colors because the
original colors sucked, and the editor sucked even more.  I've never used the
editor since, and I've rarely seen a user-supplied .rg file that had the colors
changed at all.

We'll leave the code in there for now, in case somebody whines, but I expect to
delete it eventually.



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/qt4@10004 cf6a83e7-f017-4305-b0b3-678bde24f230