* turn PluginControl into a proper widget that manages its own internal layout   on an hbox.  Instead of this crazy nonsense we used to have, we just lay out   these little hbox widgets on a grid layout, and manage all the layout that   matters to the dialog inside the dialog.  The little widget blocks manage   their own layout, which just seems a lot more sensible to manage, plus it   makes me proud that I, the layout code hater, came up with something so   reasonable, and dare I say clever.
* some kind of introductory setup for laying out into rows and columns, with no
  sensible limits imposed yet.  I'm open to suggestions on this, and I'd like to
  think of a reasonable way not to have to impose any upper limits at all.  It's
  6 columns and n rows now.  It seems like this could be balanced sensibly by
  taking the overall number of PluginControl required into account and doing
  some division or square roots or something.

TODO:

The category combo is broken somehow.  There's a comment in the code that
probably offers a clue as to why.

Settle on some kind of layout for rows and columns and upper limits.

Synth plugins are broken so far, and I haven't been able to test this layout
with a synth plugin yet at all.

Spacing needs improving somehow (QHBox::addStretch() ? ) so the knobs will be
more inclined to line up properly, and/or we might want to use SqueezedLabel
instead, and limit the label bits to some fixed width to get everything to line
up that way instead.  Small labels with a tooltip that shows the whole bit could
be a way to make it all more compact, and cram more controls into a given space.
This requires fixing SqueezedLabel, of course, which currently doesn't do
anything at all, I don't think, and certainly doesn't work right.

Overall I think it's a winner though.  This dialog was always kind of half baked
and crappy anyway, and there isn't much here that would hold up a potential
release.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/qt4@10033 cf6a83e7-f017-4305-b0b3-678bde24f230