all I'm going to be able to do until I figure out the random crash
I'm trying to work on the plugin controls not showing.  I'm not sure if the
layout keeps expanding to contain hboxes full of hidden controls or what, but I
suspect so, and that's the possibility I'm trying to test.  If that turns out to
be true, I'll move the show/hide bit out of PluginControl too, and maybe the
show bounds bit, depending.

Currently the program combo shows up, but the layout gets huge to contain a lot
of nothing.  Maybe that's why these were QObject before and had their layout
managed externally.  Oh well.  I'll sort it out once I reach a point where I can
actually test the code I'm trying to work on for at least two successive runs
without a string of random crashes in between.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/qt4@10035 cf6a83e7-f017-4305-b0b3-678bde24f230