Fixed VUMeter problem? No significant testing but meter now responds to changes in amplitude.
Lots of work, little tiny commit.
VUMeter::setLevel used to create a painter and draw the updated VU meter.
I guess by the warnings generated, this is no longer allowed in QT4 (?)
Correct mechanism if you want an immediate paint is to call repaint().
Note this should only be used if you really require an immediate repaint
(as we do here to animate the meters). If you can wait until the next
appropriate pause, use update().



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/qt4@10048 cf6a83e7-f017-4305-b0b3-678bde24f230