* remove kdeext/SqueezedLabel stub
* bring in KSqueezedTextLabel from KDE 4.2, and port to work in pure Qt as
  gui/widgets/SqueezedLabel

* fix the SPB checkbox not changing when disabled

* tinker with the rotaries some more (nothing to see here yet)

NOTES:

I experimented with using SqueezedLabel for QLabel for the plugin control title,
which was the whole point of tonight's exercise.  It looked awful, just awful.
Oh well, at least we fixed these labels elsewhere.  It was easier to port KDE's
code than to reinvent the wheel.

A reminder for later, I notice the centered pan knob in the audio IPB works, but
the one in the MIDI IPB doesn't.  That's probably a clue worth following.

After the Qt4 merge, we should move everything in kdeext/ to widgets/, and we
should think about renaming our Q classes and any lingering K classes to make it
more obvious they're really ours to maintain, and ours to hunt for in our own
docs.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/qt4@10059 cf6a83e7-f017-4305-b0b3-678bde24f230