* add the new File -> Save As Template feature
This piggybacks on File -> Save As to save a file with a new .rgt extension.
Unless you explicitly use Save As Template, you'll get either a Save As or a
Save dialog that wants to write to an .rg file, and hides any .rgt files by
default, so you won't write over a template by accident unless you're really
clumsy.  I do have a QProcess bit in there to chmod -w the file, but since we
write to a new file and then delete the original when saving, this doesn't
actually accomplish much in of itself.  My initial tests seem to indicate that
the natural level of discouragement from accidentally writing to .rgt files is
probably protection enough without going to even more elaborate lengths though,
so we'll see how it does in the field.

NOTE:  We (maybe Heikki and I, especially) should put together at least a simple
list of basic templates to offer with Thorn's release, and create a new
directory alongside examples/ and library/ for allowing access to them.  (We
also need to pull all of these things out of the resource bundle and save them
somewhere in writeable space, or at the very least somewhere in space
QFileDialog can actually access.  All of this stuff is unobtanium unless you
read files out of a source tree or a installed copy of 1.7.3 a the moment, and
we must not forget to deal with this issue before releasing.  Maybe we should
have a limited installation to $PREFIX for these things, or maybe we should just
save a copy out to ~/.local at runtime or whatever.  I don't know yet, and am
open to opinions.  I'm even more open to somebody else working this out and
getting it done.  Hint hint.)


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/qt4@10145 cf6a83e7-f017-4305-b0b3-678bde24f230