* remove conditional build ifdef gibberish for ALSA, LIBJACK, DSSI, LADSPA,   LIBRDF, XFT, LIBLO, and LIRC
NOTES: I have second thoughts about removing the ifdef for ALSA, but in for a
penny, in for a pound, and I had already done the work by the time I had the
second thoughts.

This adds the missing -llirc_client to qt4-makefile and removes the -DHAVE_x.
I didn't update the other versions of the makefile.

It looked like more if/then/maybe logic was involved with Xft, like maybe Chris
was thinking about switching to something else, or rightfully supporting
parallel options.  I probably didn't do a consistent job of just commenting out
all the ifdef and everything after the else on those files, but that's what I
was aiming to do, rather than removing the alternative "no library available"
implementations entirely in that case.

This builds and appears to work fine at a glance, but we should watch carefully
for the consequences of any careless work on my part.  This was a pretty huge
and sweeping set of changes, and while I edited cautiously, I might very well
have left some bit of "no library available" logic in parallel with the other
stuff, and I might have introduced some strange new bugs with this.  Hopefully
not, but I wanted to raise awareness of the possibility.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10187 cf6a83e7-f017-4305-b0b3-678bde24f230