* the Preview with LilyPond target actually works now, with many caveats and   much ugly crudeness, but basically as a proof of concept for rewriting a   bash/Perl script inside a dialog ctor, it's there
NOTES: Skipping all the progress monitoring nonsense makes the thing VASTLY less
complicated, though QProcess does provide the means to monitor the stream of
text, and that could be used to update the real code progress bar in a fashion
similar to the original.  I don't think it's worth it.  The "Knight Rider Busy
Bonker" looks stupid, but it's probably sufficient just to jump the progress bar
in huge stages, 25% 50% 75% 100% and be done with it.

Putting the "External Applications" thing on the General page doesn't seem quite
right, but it seems cleaner than putting these willy nilly on different tabs
where needed.  Also, we haven't yet established which external applications will
still be needed.  That remains to be seen after someone (unfortunately, probably
me) rewrites the project packager and audio file importer, attempting to convert
these to use libs instead of standalone apps as much as possible.  (I'm afraid
of introducing a .rgp version 2 format as a result of this, since the whole
thing may well come out working a little differently in the end.  This should be
avoided, but not avoided at all costs.  Especially if we can still *read* old
files.)


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/real_code_experiments@10247 cf6a83e7-f017-4305-b0b3-678bde24f230