* a day spent refactoring and and refactoring some more to solve the blocking   problem the hard way
NB: QProcess doesn't actually have a processExited(QProcess *) signal.  I just
block copied that from somewhere, and it's wrong, so all of this needs to be
rethought, and there are probably bugs elsewhere in our code too, such as the
place I block copied the connect command from.

I'll fool with it tomorrow.  Sleep beckons.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/real_code_experiments@10248 cf6a83e7-f017-4305-b0b3-678bde24f230