* experimental "Disable Permanently" button for the timer resolution warning   dialog, applied as an exercise in how to resurrect the old "don't show this   again" functionality without subclassing or replacing QMessageBox
Committing this un-tested because I can't actually test it on my main box.  I
realize disabiling this particular warning may be controversial, but OTOH users
have reported getting on fine in spite of these warnings, and my experience on
the test box is the same.  If nothing else, we'll see if this is an adequate
replacement for the old [x] Don't show this again that are not possible to add
through standard Qt classes.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10260 cf6a83e7-f017-4305-b0b3-678bde24f230