* patch from Thorsten to solve a build problem he had; details follow
"Today I tried to compile rosegarden on a rather new system and the 
configure-script asked me to install lots of missing packets. After I 
fullfilled all complaints, I still got an compile error due to some 
missing header file, so I added one line to configure.ac to test for this 
file."

I'm not exactly sure what this is about myself, as it looks like an X11 session
management lib, and I don't think we use it directly, or if we do for legacy
reasons, we might not in the long run.  For now I'm committing the patch, but
I'm not updating our build dependency docs.  I'll wait to gauge how many
"configure failed because I lack this obscure header" messages there are, and
then react accordingly.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10333 cf6a83e7-f017-4305-b0b3-678bde24f230