* more incremental work
From here it looks to get worse and worse.  The drawing and style problems are
resolved far enough to reveal that the thing is just busted for some reason or
other.

Nobody knows anything about this code, least of all me, so it pretty much sucks
from here on out.

Bigtime.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10354 cf6a83e7-f017-4305-b0b3-678bde24f230