* upon unbundling, ResourceFinder makes the resulting local user file writeable   (if this piques your interest as a potential source of controversy, read the   inline comment and then opine away)
* heavier sorting of the broken guitar chord dialog, resulting in real tangible
  progress this time, but still many bugs

Dialog doesn't come up with the default chord visible in the fingering box
doodad.  I figure that can be addressed with some slot calls.

Fret dots seem to be drawn one fret too high.

I've never seen the fret number displayed, and have a vague feeling it's being
drawn in white and/or just not showing up.

One of the combos is apparently editable, though I know not quite why.  It's
styled in orange "here I am" colors never before seen in all this work.  That
needs to be investigated, although if it proves necessary to leave that combo
editable, the style is obviously working, so it's not a major problem I don't
think.  Just a curious oddity.

Many crashes and general fragility.

Writing of user chords not tested yet.

&c.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10358 cf6a83e7-f017-4305-b0b3-678bde24f230