* patch from Immanuel Litzroth laying the foundation for the integration of the   Tranzport device into Rosegarden
"Currently only the following buttons work:
track next, track previous, mute work on the currently selected track.
rec does not seem to work on the track but on the global recording state.
fastforward, and rewind work  and move a measure ahead.
Shift fastforward and Shift rewind both move to the beginning of the
composition.
Play, Stop and record work as expected.
To be done:
* Markers & marker movement.
* Datawheel scrolling
* undo
* popping on lights.

Currently "Rosegarden" is placed in the LCD. That also needs to display
information
about the currently selected track, but I will have to look into keeping
Rosegarden
and Tranzport in sync.
Patch should be safe to use as it doesn't do anything when you don't have a
/dev/tranzport0
device  on your system."


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10380 cf6a83e7-f017-4305-b0b3-678bde24f230