* Revert to using raster renderer by default.
Although we have had problems with this so far, the main one (failure to
display anything at all in notation) should now be fixed.  Since the raster
renderer is so much faster, I think it's worth making a serious effort to
hunt and fix the remaining problems rather than switching back to native.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10428 cf6a83e7-f017-4305-b0b3-678bde24f230